arm64: dts: qcom: sc7180-el2: Add qcom,broken-reset where needed
The remoteproc PAS firmware interface does not work properly when running
bare-metal without hypervisor in EL2. It's possible to authenticate and
start new firmware, but the remoteproc is never actually brought out of
reset. On SC7180 this only affects the ADSP, MPSS works correctly.

Mark the ADSP with "qcom,broken-reset" to make the driver skip trying to
reset the remoteproc. We can still attach to it if it is already running
during boot.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
Signed-off-by: Abel Vesa <abel.vesa@oss.qualcomm.com>