drm/msm: Add soft module-dep on nvmem_qfprom
If there is an nvme phandle in the ofnode for the GPU the adreno driver
requires nvmem_qfprom to be loaded or it will fail with -EPROBE_DEFER.

Commit be6f43c64ca0 ("arm64: dts: qcom: x1e80100: Update GPU OPP table")
added a nvme phandle for the ThinkPad T14s Snapdragon X1E gen6, causing
the laptop panel to go and stay black in the initrd requiring full-disk-
encryption users to type their password blindly.

The root cause of this issue is the nvmem_qfprom not being in the initrd,
add a MODULE_SOFTDEP("pre: nvmem_qfprom") line to make initrd generators
(e.g. dracut) add the module to the initrd fixing this.

An alternative fix would be to modify various initrd generators for this,
but given that this is highly msm specific this seems like a better fix.

Fixes: be6f43c64ca0 ("arm64: dts: qcom: x1e80100: Update GPU OPP table")
Cc: Akhil P Oommen <akhilpo@oss.qualcomm.com>
Signed-off-by: Hans de Goede <johannes.goede@oss.qualcomm.com>